pyproject.toml
[project]
name = "mobot"
version = "0.1.0"
description = "Calvin's mobile robot software"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["sparkfun-pi-servo-hat", "RPi.GPIO"]
